+++ lib/molecules/dynamic_input/inputs/zwap_floating_picker/zwap_floating_picker_provider.dart
@@ -1,0 +1,32 @@
+import 'package:flutter/material.dart';
+import 'package:zwap_design_system/molecules/dynamic_input/zwap_dynamic_input.dart';
+
+class ZwapFloatingPickerProvider<T> extends ChangeNotifier {
+  final GlobalKey<ZwapDynamicInputState> inputKey;
+  final List<T> values;
+  T? _selectedValue;
+
+  final void Function(T?)? onSelected;
+
+  final String Function(T) getCopy;
+
+  T? get selectedValue => _selectedValue;
+
+  set selectedValue(T? value) => value.hashCode != _selectedValue.hashCode
+      ? () {
+          _selectedValue = value;
+          notifyListeners();
+
+          if (onSelected != null) onSelected!(value);
+        }()
+      : null;
+
+  ZwapFloatingPickerProvider({
+    required this.inputKey,
+    required this.values,
+    required this.getCopy,
+    this.onSelected,
+    T? selectedValue,
+  })  : this._selectedValue = selectedValue,
+        super();
+}